Tpetra: Fixed signed/unsigned comparison warnings in tests.
The tests concern sumIntoGlobalValues functionality for nonowned
rows in CrsMatrix.  The following build illustrates the warnings,
as well as an error which was fixed in an earlier commit.

http://testing.sandia.gov/cdash/buildSummary.php?buildid=841619

+++ packages/tpetra/test/CrsMatrix/CrsMatrix_NonlocalSumInto_Ignore.cpp
@@ -146,8 +146,8 @@ TEUCHOS_UNIT_TEST_TEMPLATE_4_DECL( CrsMatrix, NonlocalSumInto_Ignore, LocalOrdin
 
   // Get the default communicator.
   RCP<const Comm<int> > comm = Tpetra::DefaultPlatform::getDefaultPlatform ().getComm ();
-  const size_t numProcs = comm->getSize ();
-  const size_t myRank = comm->getRank ();
+  const int numProcs = comm->getSize ();
+  const int myRank = comm->getRank ();
 
   if (myRank == 0) {
     out << "Test with " << numProcs << " process" << (numProcs != 1 ? "es" : "") << endl;
